+++ egui/src/menu.rs
@@ -117,7 +117,7 @@ fn menu_impl<'c>(
 
         // TODO: this prevents sub-menus in menus. We should fix that.
         if ui.input().key_pressed(Key::Escape)
-            || ui.input().pointer.click && !button_response.clicked()
+            || ui.input().pointer.any_click() && !button_response.clicked()
         {
             bar_state.open_menu = None;
         }