+++ egui_demo_lib/src/apps/demo/drag_and_drop.rs
@@ -25,7 +25,7 @@ pub fn drag_source(ui: &mut Ui, id: Id, body: impl FnOnce(&mut Ui)) {
         // (anything with `Order::Tooltip` always gets an empty `Response`)
         // So this is fine!
 
-        if let Some(pointer_pos) = ui.input().pointer.pos {
+        if let Some(pointer_pos) = ui.input().pointer.interact_pos() {
             let delta = pointer_pos - response.rect.center();
             ui.ctx().translate_layer(layer_id, delta);
         }
@@ -126,8 +126,7 @@ impl super::View for DragAndDropDemo {
                             ui.label(item);
                         });
 
-                        let this_item_being_dragged = ui.memory().is_being_dragged(item_id);
-                        if this_item_being_dragged {
+                        if ui.memory().is_being_dragged(item_id) {
                             source_col_row = Some((col_idx, row_idx));
                         }
                     }
@@ -143,7 +142,7 @@ impl super::View for DragAndDropDemo {
 
         if let Some((source_col, source_row)) = source_col_row {
             if let Some(drop_col) = drop_col {
-                if ui.input().pointer.released {
+                if ui.input().pointer.any_released() {
                     // do the drop:
                     let item = self.columns[source_col].remove(source_row);
                     self.columns[drop_col].push(item);