+++ egui_demo_lib/src/frame_history.rs
@@ -79,7 +79,7 @@ impl FrameHistory {
         let rect = rect.shrink(4.0);
         let line_stroke = Stroke::new(1.0, Color32::from_additive_luminance(128));
 
-        if let Some(pointer_pos) = ui.input().pointer.pos {
+        if let Some(pointer_pos) = ui.input().pointer.tooltip_pos() {
             if rect.contains(pointer_pos) {
                 let y = pointer_pos.y;
                 shapes.push(Shape::line_segment(